+++ site/src/primitive.rs
@@ -273,7 +273,7 @@ fn all_uns() -> impl IntoView {
             { inverse_row([Group], No, "Inner function must be invertible", "°⊕□ {1 2_3_4 5_6}") }
             { inverse_row([Repeat], Required, "Inner function must be invertible", "°(⍥(×2)5) 1024") }
             { inverse_row([Stack], No, "", "°? 5") }
-            { inverse_row([Dump], No, "", "°dump△ [2 3 4]") }
+            { inverse_row([Dump], No, "", "°‽△ [2 3 4]") }
             { inverse_row([Pop], RequiresFill, "", "⬚5°◌") }
             { inverse_row([AudioEncode], Optional, "Decodes bytes", None) }
             { inverse_row([ImageEncode], Optional, "Decodes bytes", None) }